+++ src/onegov.ballot/HISTORY.rst
@@ -4,7 +4,7 @@ Changelog
 Unreleased
 ~~~~~~~~~~
 
-- Adds new columns (year, color, total votes) to party results.
+- Adds a status (unknown, interim, final) to elections and votes.
   [msom]
 
 1.3.2 (2017-04-27)